+++ example/android/app/src/main/java/co/ronash/pushe_example/MyFCMService.java
@@ -1,0 +1,37 @@
+package co.ronash.pushe_example;
+
+// To make this work add firebase_messaging to your flutter app.
+
+//import com.google.firebase.messaging.RemoteMessage;
+//import io.flutter.plugins.firebasemessaging.FlutterFirebaseMessagingService;
+
+@SuppressWarnings("SpellCheckingInspection")
+public class MyFCMService /* extends FlutterFirebaseMessagingService*/ {
+//    @Override
+//    public void onNewToken(String s) {
+//        co.ronash.pushe.Pushe.getFcmHandler(this).onNewToken(s);
+//        super.onNewToken(s);
+//    }
+//    @Override
+//    public void onDeletedMessages() {
+//        co.ronash.pushe.Pushe.getFcmHandler(this).onDeletedMessages();
+//        super.onDeletedMessages();
+//    }
+//    @Override
+//    public void onSendError(String s, Exception e) {
+//        co.ronash.pushe.Pushe.getFcmHandler(this).onSendError(s, e);
+//        super.onSendError(s, e);
+//    }
+//    @Override
+//    public void onMessageSent(String s) {
+//        co.ronash.pushe.Pushe.getFcmHandler(this).onMessageSent(s);
+//        super.onMessageSent(s);
+//    }
+//    @Override
+//    public void onMessageReceived(RemoteMessage remoteMessage) {
+//        if (!co.ronash.pushe.Pushe.getFcmHandler(this).onMessageReceived(remoteMessage)) {
+//            // It is for fire base, otherwise the condition will handle the message for Pushe
+//            super.onMessageReceived(remoteMessage);
+//        }
+//    }
+}